+++ packages/editor/editor-core/src/plugins/macro/actions.ts
@@ -1,21 +1,21 @@
 import { EditorState, NodeSelection } from 'prosemirror-state';
 import { Node as PmNode } from 'prosemirror-model';
 import { EditorView } from 'prosemirror-view';
-import { MacroProvider, MacroAttributes } from './types';
+import { ExtensionProvider, ExtensionAttributes } from './types';
 import { pluginKey } from './';
 import * as assert from 'assert';
 import { getValidNode } from '@atlaskit/editor-common';
 import { safeInsert, replaceSelectedNode } from 'prosemirror-utils';
 
 export const insertMacroFromMacroBrowser = (
-  macroProvider: MacroProvider,
+  extensionProvider: ExtensionProvider,
   macroNode?: PmNode,
 ) => async (view: EditorView): Promise<boolean> => {
-  if (!macroProvider) {
+  if (!extensionProvider) {
     return false;
   }
   // opens MacroBrowser for editing "macroNode" if passed in
-  const newMacro: MacroAttributes = await macroProvider.openMacroBrowser(
+  const newMacro: ExtensionAttributes = await extensionProvider.editExtension(
     macroNode,
   );
   if (newMacro) {
@@ -40,7 +40,7 @@ export const insertMacroFromMacroBrowser = (
 };
 
 export const resolveMacro = (
-  macro?: MacroAttributes,
+  macro?: ExtensionAttributes,
   state?: EditorState,
 ): PmNode | null => {
   if (!macro || !state) {
@@ -65,19 +65,20 @@ export const resolveMacro = (
   return node;
 };
 
-// gets the macroProvider from the state and tries to autoConvert a given text
+// gets the extensionProvider from the state and tries to autoConvert a given text
 export const runMacroAutoConvert = (
   state: EditorState,
   text: String,
 ): PmNode | null => {
   const macroPluginState = pluginKey.getState(state);
 
-  const macroProvider = macroPluginState && macroPluginState.macroProvider;
-  if (!macroProvider || !macroProvider.autoConvert) {
+  const extensionProvider =
+    macroPluginState && macroPluginState.extensionProvider;
+  if (!extensionProvider || !extensionProvider.autoConvert) {
     return null;
   }
 
-  const macroAttributes = macroProvider.autoConvert(text);
+  const macroAttributes = extensionProvider.autoConvert(text);
   if (!macroAttributes) {
     return null;
   }
@@ -86,21 +87,21 @@ export const runMacroAutoConvert = (
   return resolveMacro(macroAttributes, state);
 };
 
-export const setMacroProvider = (provider: Promise<MacroProvider>) => async (
-  view: EditorView,
-): Promise<boolean> => {
-  let resolvedProvider: MacroProvider | null;
+export const setExtensionProvider = (
+  provider: Promise<ExtensionProvider>,
+) => async (view: EditorView): Promise<boolean> => {
+  let resolvedProvider: ExtensionProvider | null;
   try {
     resolvedProvider = await provider;
     assert(
-      resolvedProvider && resolvedProvider.openMacroBrowser,
-      `MacroProvider promise did not resolve to a valid instance of MacroProvider - ${resolvedProvider}`,
+      resolvedProvider && resolvedProvider.editExtension,
+      `ExtensionProvider promise did not resolve to a valid instance of ExtensionProvider - ${resolvedProvider}`,
     );
   } catch (err) {
     resolvedProvider = null;
   }
   view.dispatch(
-    view.state.tr.setMeta(pluginKey, { macroProvider: resolvedProvider }),
+    view.state.tr.setMeta(pluginKey, { extensionProvider: resolvedProvider }),
   );
   return true;
 };