+++ packages/elements/task-decision/src/types.ts
@@ -1,5 +1,4 @@
 import { ServiceConfig } from '@atlaskit/util-service-support';
-import { PubSubClient } from '@atlassian/pubsub';
 
 export type DecisionState = 'DECIDED';
 export type DecisionStatus = 'CREATED';
@@ -226,3 +225,30 @@ export interface OnUpdate<T> {
 }
 
 export type Appearance = 'inline' | 'card';
+
+/**
+ * Same as PubSub client types (don't want a direct dep though)
+ */
+
+export type ARI = string;
+export type AVI = string;
+
+export interface PubSubOnEvent<T = any> {
+  (event: string, data: T): void;
+}
+
+export interface PubSubClient {
+  on(eventAvi: string, listener: PubSubOnEvent): PubSubClient;
+
+  off(eventAvi: string, listener: PubSubOnEvent): PubSubClient;
+
+  join(aris: ARI[]): Promise<PubSubClient>;
+
+  leave(aris: ARI[]): Promise<PubSubClient>;
+}
+
+export enum PubSubSpecialEventType {
+  ERROR = 'ERROR',
+  CONNECTED = 'CONNECTED',
+  RECONNECT = 'RECONNECT',
+}