+++ packages/editor/editor-markdown-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-markdown-transformer
 
+## 0.2.13
+
 ## 0.2.12
 
 ## 0.2.11